requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.23
pydantic==2.4.2
pydantic-settings==2.0.3
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
email-validator==2.1.0.post1
mysql-connector-python==8.2.0
alembic==1.12.1
redis==4.6.0
fastapi-limiter==0.1.5
python-dotenv==1.0.0
jinja2==3.1.2
aiosmtplib==2.0.2 